[9.x] Move reusable `onNotSuccessfulTest` to `TestResponse`
This allows it to be reusable on Testbench without having redundant code.
